refactor: showConnectedDevices
